PerfTests: rework perftest organization a bit
