+++ core/perf_test/PerfTestDriver.hpp
@@ -398,91 +398,5 @@ void run_test_mdrange( int exp_beg , int exp_end, const char deviceTypeName[], i
 }
 
 
-template< class DeviceType >
-void run_test_hexgrad( int exp_beg , int exp_end, const char deviceTypeName[] )
-{
-  std::string label_hexgrad ;
-  label_hexgrad.append( "\"HexGrad< double , " );
-  // mfh 06 Jun 2013: This only appends "DeviceType" (literally) to
-  // the string, not the actual name of the device type.  Thus, I've
-  // modified the function to take the name of the device type.
-  //
-  //label_hexgrad.append( KOKKOS_MACRO_TO_STRING( DeviceType ) );
-  label_hexgrad.append( deviceTypeName );
-  label_hexgrad.append( " >\"" );
-
-  for (int i = exp_beg ; i < exp_end ; ++i) {
-    double min_seconds = 0.0 ;
-    double max_seconds = 0.0 ;
-    double avg_seconds = 0.0 ;
-
-    const int parallel_work_length = 1<<i;
-
-    for ( int j = 0 ; j < NUMBER_OF_TRIALS ; ++j ) {
-      const double seconds = HexGrad< DeviceType >::test(parallel_work_length) ;
-
-      if ( 0 == j ) {
-        min_seconds = seconds ;
-        max_seconds = seconds ;
-      }
-      else {
-        if ( seconds < min_seconds ) min_seconds = seconds ;
-        if ( seconds > max_seconds ) max_seconds = seconds ;
-      }
-      avg_seconds += seconds ;
-    }
-    avg_seconds /= NUMBER_OF_TRIALS ;
-
-    std::cout << label_hexgrad
-      << " , " << parallel_work_length
-      << " , " << min_seconds
-      << " , " << ( min_seconds / parallel_work_length )
-      << std::endl ;
-  }
-}
-
-template< class DeviceType >
-void run_test_gramschmidt( int exp_beg , int exp_end, const char deviceTypeName[] )
-{
-  std::string label_gramschmidt ;
-  label_gramschmidt.append( "\"GramSchmidt< double , " );
-  // mfh 06 Jun 2013: This only appends "DeviceType" (literally) to
-  // the string, not the actual name of the device type.  Thus, I've
-  // modified the function to take the name of the device type.
-  //
-  //label_gramschmidt.append( KOKKOS_MACRO_TO_STRING( DeviceType ) );
-  label_gramschmidt.append( deviceTypeName );
-  label_gramschmidt.append( " >\"" );
-
-  for (int i = exp_beg ; i < exp_end ; ++i) {
-    double min_seconds = 0.0 ;
-    double max_seconds = 0.0 ;
-    double avg_seconds = 0.0 ;
-
-    const int parallel_work_length = 1<<i;
-
-    for ( int j = 0 ; j < NUMBER_OF_TRIALS ; ++j ) {
-      const double seconds = ModifiedGramSchmidt< double , DeviceType >::test(parallel_work_length, 32 ) ;
-
-      if ( 0 == j ) {
-        min_seconds = seconds ;
-        max_seconds = seconds ;
-      }
-      else {
-        if ( seconds < min_seconds ) min_seconds = seconds ;
-        if ( seconds > max_seconds ) max_seconds = seconds ;
-      }
-      avg_seconds += seconds ;
-    }
-    avg_seconds /= NUMBER_OF_TRIALS ;
-
-    std::cout << label_gramschmidt
-      << " , " << parallel_work_length
-      << " , " << min_seconds
-      << " , " << ( min_seconds / parallel_work_length )
-      << std::endl ;
-  }
-}
-
 }
