+++ core/perf_test/PerfTest_Category.hpp
@@ -1,13 +1,13 @@
 /*
 //@HEADER
 // ************************************************************************
-// 
+//
 //                        Kokkos v. 2.0
 //              Copyright (2014) Sandia Corporation
-// 
+//
 // Under the terms of Contract DE-AC04-94AL85000 with Sandia Corporation,
 // the U.S. Government retains certain rights in this software.
-// 
+//
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions are
 // met:
@@ -36,20 +36,30 @@
 // SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 //
 // Questions? Contact  H. Carter Edwards (hcedwar@sandia.gov)
-// 
+//
 // ************************************************************************
 //@HEADER
 */
 
+#ifndef KOKKOS_TEST_THREADS_HPP
+#define KOKKOS_TEST_THREADS_HPP
+
 #include <gtest/gtest.h>
-#include <Kokkos_Core.hpp>
 
-int main(int argc, char *argv[]) {
-  ::testing::InitGoogleTest(&argc,argv);
-  Kokkos::initialize(argc,argv);
+namespace Test {
 
-  int result = RUN_ALL_TESTS();
+class default_exec : public ::testing::Test {
+protected:
+  static void SetUpTestCase() {
+  }
 
-  Kokkos::finalize();
-  return result;
-}
+  static void TearDownTestCase() {
+  }
+};
+
+} // namespace Test
+
+#define TEST_CATEGORY default_exec
+#define TEST_EXECSPACE Kokkos::DefaultExecutionSpace
+
+#endif
